Added blended bspline enhancement, used to mdoify the behavior at the ends of the spline, in order to match James Franklin's heuristic smoothing scheme employed in his Editsonde program.
The use may specify, for either or both ends of the series, a internval over which the spline results are linearly blended with the original observations. The blend weights the original observation at 100% at the end of the series, and decreases to 0% at the distance of the specified blending interval.

For slope, the original slopes are computed as finite difference (since this is what James uses), the belnding mixes these with the spline derived slopes. The finite difference derived slopes are from centered differences for all poits except the ends. A single sided difference is used for each end point finite difference slope.


git-svn-id: http://svn.eol.ucar.edu/svn/eol/bspline/trunk@6410 da89c675-a4d7-0310-8a15-9e7546503515