remove packaging and rcsid version
Packaging is obsolete when the source can just be
downloaded from github, and rcsid stopped working
long ago.  Replace it with gitinfo from eol_scons,
but provide a default version in a header file so
source can still be used as a header-only library.
Also, replace ImplVersion() and IfaceVersion() with just Version().