add missing assignment operator, use unique_ptr
replace raw private base pointer with unique_ptr to get automatic
deletion on destruction, and to get compiler warnings if
an implicit method ever tries to copy it.  This is the better way to
implement the pimpl idiom.  I did not add move operators because
I don't think they are really warranted.  Fixes issue #3.